requirements.txt
# Основные зависимости
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0
streamlit>=1.28.0

# OCR библиотеки
paddlepaddle>=3.0.0
paddleocr>=3.1.0
easyocr>=1.7.0
pytesseract>=0.3.10

# Перевод - используем deep-translator вместо googletrans
deep-translator>=1.11.4

# Базовые утилиты
pyyaml>=6.0
requests>=2.28.0

# Обработка изображений
scikit-image>=0.19.0